Completed "Task #18 -- Implement the CC0 machine interpreter for J1"
In the course of doing this, which first seemed difficult before
seeming easier than expected, I realized that my small-step interpreter
was all wrong right down to find_redex() , so I reimplemented
find_redex() and the small-step interpreter.

Now I think I actually understand what it was supposed to do in the
first place, but at least now my code now works in the second place,
which is really the only place that matters as it is the final place.

Signed-off-by: Joel Savitz <joelsavitz@gmail.com>