Completed "Task #22 - Implement the CK0 machine interpreter for J1"
wew

I didn't have to spend that much time fixing trivial mistakes, but
there was one bug in olist_insert() that took me a while to spot,
which was a mishandling of the push_front case needed for my stack.
That is now fixed and the CK0 machine interpreter works as specified,
as far as I know.

Signed-off-by: Joel Savitz <joelsavitz@gmail.com>